Fix registration of KeyTool in the bootloader
This accidentally used the complete path, rather than the relative
path in the ESP. Also get rid of machine-id, as that makes the boot
loader default to KeyTool, which is not what we want.

It's questionable to install stuff by machine-id, but I think it
has the benefit that we do not accidentally overwrite a KeyTool
signed by another key that is already in use.